BAccordion script setup update
